Measure compression statistics per remote client addr.
Compression statistics, like number of sent bytes are now grouped in RemoteCompressionClient class. This class is saved in a HashMap where key is the remote client address.

The RemoteCompressionClient is created in init_qcmp and is removed after inactivity longer than 60 minutes for now.

Hopefully `baseRequest.getRemoteAddr();` is good enaugh key for now.